requirements.txt
# Python dependencies for the ear training application
# You can install these with pip using:
#     pip install -r requirements.txt

# Qt bindings for the GUI
pyside6>=6.5

# Audio playback and recording
sounddevice>=0.5
soundfile>=0.12

# Musical data handling and analysis
music21>=8.1
mido[ports-rtmidi]>=1.3
python-rtmidi>=1.5

# Pitch detection
#aubio>=0.4

# Optional: fluid synthesizer for realistic instrument sounds
pyfluidsynth>=1.3

# Numerical and plotting utilities
numpy>=1.25
matplotlib>=3.8
pretty_midi>=0.2

# For processing HDF5 metadata (Million Song Dataset)
h5py>=3.8